Fixed a bug where a shelf deletion wasn't propagating to links
Since SQLite by default doesn't enable foreign keys, whenever a shelf
was deleted their links would still be in place and pointing to the ID
of the shelf.

This together with the fact that SQLite reuses IDs meant that when a new
shelf was created it would have the contents of a previously deleted
one.

This issue was fixed by simply enabling foreign keys every time a new
connection is made to the database.

Fixes #8